fix in preparation for rust 1.3

+++ src/data/mod.rs
@@ -121,7 +121,7 @@ macro_rules! cnst {
             fn generate(&self, dims: Dim4) -> Array {
                 unsafe {
                     let mut temp: i64 = 0;
-                    af_constant(&mut temp as MutAfArray, *self as c_double,
+                    af_constant(&mut temp as MutAfArray, *self as u64 as c_double,
                                 dims.ndims() as c_uint, dims.get().as_ptr() as *const DimT,
                                 $ffi_type);
                     Array::from(temp)
